gpu_example_code_readonly/requirements.txt
cupy-cuda11x>=12.0.0     # choose the build matching your CUDA toolkit
numpy>=1.23.0
scipy>=1.9.0
ray>=2.0.0
pandas>=1.3.0
matplotlib>=3.4.0